requirements.txt
# Core
flask==3.1.3
requests==2.33.1
python-dotenv==1.2.2
PyYAML==6.0.3

# Data & Knowledge
chromadb==1.5.5
pydantic==2.12.5

# AI & LLM
google-genai==1.70.0
google-generativeai==0.8.6

# Utilities
rich==14.3.3
tqdm==4.67.3
